Fixed UI and datatypes

+++ lib/components/add_exercise_dialog.dart
@@ -7,7 +7,7 @@ class AddExerciseDialog extends StatelessWidget {
   TextEditingController repsController = TextEditingController();
   TextEditingController weightController = TextEditingController();
   
-  final Function(String, int, int, int) onAdd;
+  final Function(String, int, int, double) onAdd;
 
   @override
   Widget build(BuildContext context) {
@@ -73,7 +73,7 @@ class AddExerciseDialog extends StatelessWidget {
                         String workoutName = nameController.text;
                         int? reps = int.tryParse(repsController.text);
                         int? sets = int.tryParse(setsController.text);
-                        int? weight = int.tryParse(weightController.text);
+                        double? weight = double.tryParse(weightController.text);
                         if(workoutName.isNotEmpty && reps != null && sets != null && weight != null) {
               onAdd(workoutName, reps, sets, weight);
                         }
